duh, again.
svn:sync-xref-src-repo-rev: r77423

+++ src/VBox/HostDrivers/Support/SUPDrv.d
@@ -1,4 +1,4 @@
-/* $Id: SUPDrv.d 40607 2012-03-24 21:48:27Z knut.osmundsen@oracle.com $ */
+/* $Id: SUPDrv.d 40881 2012-04-11 23:33:21Z knut.osmundsen@oracle.com $ */
 /** @file
  * SUPDrv - Static dtrace probes.
  */
@@ -20,8 +20,8 @@ provider vboxdrv
 {
     probe session__create(struct SUPDRVSESSION *pSession, int fUser);
     probe session__close(struct SUPDRVSESSION *pSession);
-    probe ioctl__entry(struct SUPDRVSESSION *pSession, unsigned int uIOCtl, void *pvReqHdr);
-    probe ioctl__return(struct SUPDRVSESSION *pSession, unsigned int uIOCtl, void *pvReqHdr, int rc, int rcReq);
+    probe ioctl__entry(struct SUPDRVSESSION *pSession, uintptr_t uIOCtl, void *pvReqHdr);
+    probe ioctl__return(struct SUPDRVSESSION *pSession, uintptr_t uIOCtl, void *pvReqHdr, int rc, int rcReq);
 };
 
 #pragma D attributes Evolving/Evolving/Common provider vboxdrv provider
